Move ItemSelector component to a separate folder
