+++ src/components/_pages/logging-settings/EventLogsSettingsForm/index.tsx
@@ -1,5 +1,5 @@
-import ItemSelector from 'components/_pages/logging-settings/ItemSelector';
 import CheckboxField from 'components/Input/CheckboxField';
+import ItemSelector from 'components/ItemSelector';
 import ProgressButton from 'components/ProgressButton';
 import Widget from 'components/Widget';
 import { selectors as enumSelectors, getEnumLabel } from 'ducks/enums';
@@ -87,7 +87,7 @@ const EventLogsSettingForm = () => {
                     value: el,
                 }))
                 .sort((a, b) => a.label.localeCompare(b.label)),
-        [moduleEnum],
+        [resourceEnum],
     );
 
     return (
